add UsePrecompiledHeader flag and change vendor projects to Utility type

+++ sharpmake/src/Common/BaseProject.cs
@@ -6,6 +6,7 @@ using System.IO;
 public abstract class BaseProject : Project
 {
     public string TypeNameWithoutPostfix => GetType().Name.Replace("Project", "").ToLowerInvariant();
+    public bool UsePrecompiledHeaders = true;
 
     protected BaseProject()
     {
@@ -47,8 +48,11 @@ public abstract class BaseProject : Project
 
         if (BaseConfiguration.IsCPP)
         {
-            conf.PrecompHeader = conf.PrecompHeader == null ? $"{Name}_stdafx.hpp" : conf.PrecompHeader;
-            conf.PrecompSource = conf.PrecompSource == null ? $"{Name}_stdafx.cpp" : conf.PrecompSource;
+            if (UsePrecompiledHeaders)
+            {
+                conf.PrecompHeader = conf.PrecompHeader == null ? $"{Name}_stdafx.hpp" : conf.PrecompHeader;
+                conf.PrecompSource = conf.PrecompSource == null ? $"{Name}_stdafx.cpp" : conf.PrecompSource;
+            }
 
             var absoluteSourceRootPath = Path.GetFullPath(SourceRootPath);
             if (!Directory.Exists(absoluteSourceRootPath))
@@ -57,13 +61,13 @@ public abstract class BaseProject : Project
             }
 
             var precompHeaderPath = Path.Combine(absoluteSourceRootPath, conf.PrecompHeader);
-            if (!File.Exists(precompHeaderPath))
+            if (!File.Exists(precompHeaderPath) && UsePrecompiledHeaders)
             {
                 File.WriteAllText(precompHeaderPath, $"#pragma once{Environment.NewLine}");
             }
 
             var precompSourcePath = Path.Combine(absoluteSourceRootPath, conf.PrecompSource);
-            if (!File.Exists(precompSourcePath))
+            if (!File.Exists(precompSourcePath) && UsePrecompiledHeaders)
             {
                 File.WriteAllText(precompSourcePath, $"#include \"{conf.PrecompHeader}\"{Environment.NewLine}");
             }
